+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Solidity prettierrc
 - announce testament execution
 - decline testament execution
+- execute testament
 
 ### Removed